Remove margin and padding resets from  the universal selector and set it on the nav
Remove margin:0 and padding:0 in the universal selector so as not to
offset margin and padding on every element in the body.
The margin and padding resets were set on the ul in the nav.
Only margin:0 was set on the logo heading to remove default
heading margin and align the logo with the list items.
Margin top was added on the logo to vertically align elements in the nav.